// thunder-sdk/src/types/transfer_managers/transfer_manager_721/TransferManager721Abi.d.ts
/* Autogenerated file. Do not edit manually. */

/* tslint:disable */
/* eslint-disable */

/*
  Fuels version: 0.37.0
  Forc version: 0.35.3
  Fuel-Core version: 0.17.3
*/

import type {
  BigNumberish,
  BN,
  BytesLike,
  Contract,
  DecodedValue,
  FunctionFragment,
  Interface,
  InvokeFunction,
} from 'fuels';

import type { Enum } from "./common";

export type IdentityInput = Enum<{ Address: AddressInput, ContractId: ContractIdInput }>;
export type IdentityOutput = Enum<{ Address: AddressOutput, ContractId: ContractIdOutput }>;

export type AddressInput = { value: string };
export type AddressOutput = AddressInput;
export type ContractIdInput = { value: string };
export type ContractIdOutput = ContractIdInput;

interface TransferManager721AbiInterface extends Interface {
  functions: {
    get_exchange: FunctionFragment;
    initialize: FunctionFragment;
    transfer_nft: FunctionFragment;
  };

  encodeFunctionData(functionFragment: 'get_exchange', values: []): Uint8Array;
  encodeFunctionData(functionFragment: 'initialize', values: [ContractIdInput]): Uint8Array;
  encodeFunctionData(functionFragment: 'transfer_nft', values: [ContractIdInput, IdentityInput, IdentityInput, BigNumberish, BigNumberish]): Uint8Array;

  decodeFunctionData(functionFragment: 'get_exchange', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'initialize', data: BytesLike): DecodedValue;
  decodeFunctionData(functionFragment: 'transfer_nft', data: BytesLike): DecodedValue;
}

export class TransferManager721Abi extends Contract {
  interface: TransferManager721AbiInterface;
  functions: {
    get_exchange: InvokeFunction<[], ContractIdOutput>;
    initialize: InvokeFunction<[exchange_contract: ContractIdInput], void>;
    transfer_nft: InvokeFunction<[collection: ContractIdInput, from: IdentityInput, to: IdentityInput, token_id: BigNumberish, amount: BigNumberish], void>;
  };
}